git-svn-id: http://deepneko.dyndns.org/svn/repos/explorer@233 76beb010-293a-44e9-8231-3d3008485057

+++ bin/encode.rb
@@ -29,7 +29,7 @@ movie_option = " and (path like '%.avi' or path like '%.wmv')"
 if getopt[:a]
   encodelist = $con.execute("select path from filelist where flv=''" + movie_option).flatten
 elsif getopt[:d]
-  encodelist = $con.execute("select path from filelist where path like '#{getopt[:d]}%' and flv=''" + movie_option).flatten
+  encodelist = $con.execute("select path, flv from filelist where path like '#{getopt[:d]}%'" + movie_option).flatten
 elsif getopt[:f]
   encodelist = $con.execute("select path from filelist where path like '%#{getopt[:f]}'" + movie_option).flatten
 end
@@ -37,26 +37,28 @@ end
 # 1. scp avi,wmv,mpg local2remote
 # 2. ffmpeg encode at remote host
 # 3. scp flv remote2local
-encodelist.each do |path|
+encodelist.each do |path, flv|
   src = File.basename(path)
   dist = Digest::MD5.new.update(src).to_s + ".flv"
 
-  # command
-  scp_up = "scp -P #{$const.SSH_PORT} \"#{path}\" #{$const.ENCODE_SERVER}:~/"
-  encode = "ssh -p #{$const.SSH_PORT} #{$const.ENCODE_SERVER} '" + encode(src, dist) + "'"
-  scp_down = "scp -P #{$const.SSH_PORT} #{$const.ENCODE_SERVER}:~/#{dist} #{$const.FLV_DIRECTORY}"
-  rm = "ssh -p #{$const.SSH_PORT} #{$const.ENCODE_SERVER} 'rm -f \"#{src}\";rm -f #{dist}'"
+  if flv == dist
+    # command
+    scp_up = "scp -P #{$const.SSH_PORT} \"#{path}\" #{$const.ENCODE_SERVER}:~/"
+    encode = "ssh -p #{$const.SSH_PORT} #{$const.ENCODE_SERVER} '" + encode(src, dist) + "'"
+    scp_down = "scp -P #{$const.SSH_PORT} #{$const.ENCODE_SERVER}:~/#{dist} #{$const.FLV_DIRECTORY}"
+    rm = "ssh -p #{$const.SSH_PORT} #{$const.ENCODE_SERVER} 'rm -f \"#{src}\";rm -f #{dist}'"
 
-  # exec command
-  `#{scp_up}`
-  `#{encode}`
-  `#{scp_down}`
-  `#{rm}`
+    # exec command
+    `#{scp_up}`
+    `#{encode}`
+    `#{scp_down}`
+    `#{rm}`
 
-  begin
-    $con.execute("update filelist set flv='#{dist}' where path=\"#{path}\"")
-  rescue SQLite3::SQLException
-    p "Exception:" + dist + ":" + path + "\n"
+    begin
+      $con.execute("update filelist set flv='#{dist}' where path=\"#{path}\"")
+    rescue SQLite3::SQLException
+      p "Exception:" + dist + ":" + path + "\n"
+    end
   end
 end
